+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * add `+nan.0` and `-nan.0` (R7RS)
 * properly handle negative inexact zero
 * new `environment?` function
+* add `current-directory` and `set-...` from SRFI-170
 ### Bugfix
 * fix prism highlighting of names (for new context help)
 * fix error when using help in node REPL and there are no doc string